Cache: fix permissions getting reset every day
Having no expiration for GroupLastSync was causing a bug where we
cached groups one day and the groups were gone the day after.

This is because when we ask Okta for groups we ask only for the
groups that changed from the last sync, and groups have an expiration of
24 hours.

With this solution GroupLastSync will be deleted once a day (or
on new deploys), making the next sync fetch all groups.